third_party/blink/web_tests/external/wpt/css/css-values/calc-in-media-queries-with-mixed-units-expected.txt

This is a testharness.js-based test.
[PASS] box should be orange if the calc between px-em in @media was correct
[PASS] box should be orange if the calc between vh+em in @media was correct
[PASS] box should be orange if the calc between vw-em in @media was correct
[PASS] box should be orange if the calc between vw+vh in @media was correct
[PASS] box should be orange if the calc between vh+px in @media was correct
[PASS] box should be orange if the calc between vw+px in @media was correct
[FAIL] box should be orange if the calc between px/em*em in @media was correct
  assert_equals: expected "rgb(255, 165, 0)" but got "rgb(0, 0, 255)"
[FAIL] box should be orange if the calc between vh*em in @media was correct
  assert_equals: expected "rgb(255, 165, 0)" but got "rgb(0, 0, 255)"
[FAIL] box should be orange if the calc between vh*vw/em*px/vh in @media was correct
  assert_equals: expected "rgb(255, 165, 0)" but got "rgb(0, 0, 255)"
[FAIL] box should be orange if the calc between vw/px*vh in @media was correct
  assert_equals: expected "rgb(255, 165, 0)" but got "rgb(0, 0, 255)"
[FAIL] box should be orange if the calc between vh*vw/em*px in @media was correct
  assert_equals: expected "rgb(255, 165, 0)" but got "rgb(0, 0, 255)"
[FAIL] box should be orange if the calc between vw*vh*px*em/px/px/px in @media was correct
  assert_equals: expected "rgb(255, 165, 0)" but got "rgb(0, 0, 255)"
Harness: the test ran to completion.
